Remove useMenuRadioItem hook
